refactor(web): clean up appeal detail mappers
