+++ appeals/web/src/server/appeals/appeal-details/case-summary/case-summary.mapper.js
@@ -1,0 +1,19 @@
+import { removeSummaryListActions } from '#lib/mappers/mapper-utilities.js';
+import { isDefined } from '#lib/ts-utilities.js';
+
+/**
+ * @param {{ appeal: MappedInstructions }} mappedData
+ * @returns {PageComponent}
+ */
+export const generateCaseSummary = (mappedData) => {
+	return {
+		type: 'summary-list',
+		parameters: {
+			rows: [
+				removeSummaryListActions(mappedData.appeal.siteAddress.display.summaryListItem),
+				removeSummaryListActions(mappedData.appeal.localPlanningAuthority.display.summaryListItem)
+			].filter(isDefined),
+			classes: 'govuk-summary-list--no-border'
+		}
+	};
+};
